refactor: move LoadSurveyListSpy to a separate folder
