+++ src/presentation/test/index.ts
@@ -1,4 +1,5 @@
 export * from './mock-validation';
+export * from './mock-survey-list';
 export * from './mock-authentication';
 export * from './mock-add-account';
 export * as Helper from './form-helper';
