+++ src/presentation/test/mock-survey-list.ts
@@ -1,0 +1,13 @@
+import { mockSurveyListModel } from '@/domain/test';
+import { LoadSurveyList } from '@/domain/usecases';
+
+export class LoadSurveyListSpy implements LoadSurveyList {
+  callsCount = 0;
+  surveys = mockSurveyListModel();
+
+  async loadAll() {
+    this.callsCount++;
+
+    return this.surveys;
+  }
+}
